refactor: use button components
