+++ src/components/WordleSection.tsx
@@ -1,37 +1,18 @@
-import { useSelector, useDispatch } from 'react-redux';
-
-import { resetGuessedWord } from '../features/guessedWordSlice';
-import { resetSelection } from '../features/selectedInputSlice';
-import { resetResult } from '../features/resultSlice';
-import { updateResult } from '../features/resultSlice';
-
 import './WordleSection.css';
 
 import WordleTable from './WordleTable';
+import OkButton from './OkButton';
+import ResetButton from './ResetButton';
 import ResultDiv from './ResultDiv';
 
-import type { RootState } from '../app/store';
-
 export default function WordleSection() {   
-    const guessedWord = useSelector((state: RootState) => state.guessedWord);
-    const dispatch = useDispatch();
-    
-    function handleOk() {
-        dispatch(updateResult(guessedWord));
-    }
-
-    function handleReset() {
-        dispatch(resetGuessedWord());
-        dispatch(resetSelection());
-        dispatch(resetResult());
-    }
 
     return (
         <section>
             <WordleTable />
             <div className="button_container">
-                <button onClick={handleOk}>OK</button>
-                <button onClick={handleReset} >Reset</button>
+                <OkButton />
+                <ResetButton />
             </div>
             <ResultDiv />
         </section>
